feat&refactor: board
Genre.tsx->Board.tsx
Board.tsx->PaintBoard.tsx
Card.tsx->PaintCard.tsx
