+++ src/routes/components_mylikes/PaintCard.tsx
@@ -1,15 +1,20 @@
 import styled from "styled-components";
 import { Draggable } from "react-beautiful-dnd";
-import { ILike } from "./atoms_mylikes";
+import {
+  ILike,
+  myLikesCategoryAtom,
+  myLikesTemplateAtom,
+} from "./atoms_mylikes";
 import { useState } from "react";
+import { useRecoilValue } from "recoil";
 
 interface CardProps {
   key: string;
-  song: ILike;
+  like: ILike;
   index: number;
 }
 
-const DraggableCard = styled.div<{ song: ILike }>`
+const DraggableCard = styled.div<{ like: ILike }>`
   background-color: white;
   box-shadow: inset 0px 0px 4px 1px rgba(0,0,0,0.7);
   border-radius: 20px;
@@ -27,7 +32,9 @@ const DraggableCard = styled.div<{ song: ILike }>`
   }
 `;
 
-function Card({ song, index }: CardProps) {
+function PaintCard({ like, index }: CardProps) {
+  const currentCategory = useRecoilValue(myLikesCategoryAtom);
+  const myLikesTemplate = useRecoilValue(myLikesTemplateAtom);
   const [mouseOn, setMouseOn] = useState(false);
   const onMouseEnter = () => {
     setMouseOn(true);
@@ -36,21 +43,27 @@ function Card({ song, index }: CardProps) {
     setMouseOn(false);
   };
   return (
-    <Draggable draggableId={song.id} index={index}>
+    <Draggable draggableId={like.id} index={index}>
       {(provided) => (
         <DraggableCard
           onMouseEnter={onMouseEnter}
           onMouseLeave={onMouseLeave}
-          song={song}
+          like={like}
           ref={provided.innerRef}
           {...provided.dragHandleProps}
           {...provided.draggableProps}
         >
-          {mouseOn ? song.singer : song.title}
+          {mouseOn
+            ? myLikesTemplate[currentCategory]?.typingAttrs
+                .split(",")
+                .slice(1)
+                .map((attr) => `${like[attr]}`)
+                .join(" || ")
+            : like.title}
         </DraggableCard>
       )}
     </Draggable>
   );
 }
 
-export default Card;
+export default PaintCard;
